Move to /channel purge
